requirements.txt
blinker==1.6.2
click==8.1.6
colorama==0.4.6
dnspython==2.4.2
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JSON==0.4.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
numpy==1.26.0
pandas==2.1.1
pymongo==4.5.0
python-bsonjs==0.3.0
python-dateutil==2.8.2
pytz==2023.3.post1
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
threadpoolctl==3.2.0
tzdata==2023.3
Werkzeug==3.0.0
